backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langgraph
langchain-community
langchain-groq
langchain-chroma
langchain-ollama
langchain-text-splitters

